Expanded TraP-trans-tools
This branch was created primarily for the CHILES transient project.

The scripts now save more information such as eta and nu variable candidates, and summaries of the transient candidates. Also added numpy masks to make the filtering a bit more efficient. This was all in aid of making the scripts friendly for automated analysis of the data, and providing all the data the user might need to look further into the TraP results.